webdriverio: refactor find-strategy method
